escucha todos los comandos desde el front

+++ src/socket.h
@@ -12,7 +12,7 @@ int socket_init(socket_t* self);
 int socket_release(socket_t* self);
 int socket_connect(socket_t* self, char* address, char* service);
 int socket_listen(socket_t* self, char* service);  
-int socket_read(int client_fd, char* buffer, size_t size);
+int socket_read(int client_fd, char* buffer, int size);
 int socket_accept(socket_t* self, int* client_fd, char* service);
 int socket_send(socket_t* self, const char* buffer, int length);
